+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/tables/TablesSidebarVpsController.java
@@ -607,7 +607,7 @@ public class TablesSidebarVpsController implements Initializable, AutoCompleteTe
     List<VpsTable> tables = client.getVpsService().getTables();
     refreshSheetData(tables);
     TreeSet<String> collect = new TreeSet<>(tables.stream().map(t -> t.getDisplayName()).collect(Collectors.toSet()));
-    autoCompleteNameField = new AutoCompleteTextField(this.nameField, this, collect);
+    autoCompleteNameField = new AutoCompleteTextField(null, this.nameField, this, collect);
 
     preferencesChanged(PreferenceNames.UI_SETTINGS, null);
     client.getPreferenceService().addListener(this);