+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/tournaments/TournamentDialogs.java
@@ -34,7 +34,7 @@ public class TournamentDialogs {
   public static TournamentTable openTableSelectionDialog(Stage parent, Tournament tournament, TournamentTable tournamentTable) {
     Stage stage = Dialogs.createStudioDialogStage(parent, TournamentTableSelectorDialogController.class, "dialog-tournament-table-selector.fxml", "Tournament - Table Selection");
     TournamentTableSelectorDialogController controller = (TournamentTableSelectorDialogController) stage.getUserData();
-    controller.setTournamentTable(tournament, tournamentTable);
+    controller.setTournamentTable(stage, tournament, tournamentTable);
 
     stage.showAndWait();