+++ RELEASE_NOTES.md
@@ -7,5 +7,6 @@
 
 - **VPin Mania**: The actual highscore timestamp is now used when a highscore is written. So when you synchronize your highscores, the actual creation date will be used instead of the date of submission. This way, the "Recent Highscores" widget shows only the actual recently created highscores.
 - **VPin Mania**: Fixed a bunch of sorting and navigation issues.
-- **VPin Mania**: Added table search.
+- **VPin Mania**: Added VPS table search.
+- **VPS Table Search Input**: Fixed input fields for VPS table searches so that whitespaces are allowed now too.
 - **Mac OS**: Fixed wrong OS name detection when editing files.